Uploaded code for using voice control for a rover
This is the code to control a rover using only voice recognition commands. This code will be part of my final project to build out a prototype of the Mars Rover, in tribute to Spirit, Opportunity, Curiosity and all the other rovers that are paving a path for us to Mars.