Created realistic time lag of 1.7 seconds between command and acceptance
Mars is 17 light minutes away from Earth and so every command and accept sequence takes a 34-minute round trip. I assimilated this delay into the rover as a gameplay element for controllers to appreciate the complexities in issuing a command and factoring in a delay before a response, which adds to the excitement and appreciation of the experience.